refactor(viewer): extract the schedule presentation into ListScheduleTable
The pivot header, virtualized body and totals footer were inlined in
ListResultsTable, which the review flagged for compounding an already
oversized module (503 lines before this PR, 647 after).

They move to `ListScheduleTable` along with the column/width derivation that
only the pivot view uses. The virtualizer stays with the parent — it scrolls
the shared container and its row count switches between the two views — and
arrives as a prop.

ListResultsTable is now 547 lines. That is still over the ~400-line house
rule: the remainder is the pre-existing nested table, not the schedule view.
This PR now adds 44 lines to that file rather than 144.

Raised in the PR #1867 review. Behaviour is unchanged: viewer suite 1860
pass / 0 fail, tsc clean.
